error callback data parameter type now uint16_t (breaking compatibility)

+++ examples/ARDUINO/Local/ThroughSerial/HC-12-SendAndReceive/Transmitter/Transmitter.ino
@@ -36,7 +36,7 @@ void setup() {
 };
 
 
-void error_handler(uint8_t code, uint8_t data, void *custom_pointer) {
+void error_handler(uint8_t code, uint16_t data, void *custom_pointer) {
   if(code == PJON_CONNECTION_LOST) {
     Serial.print("Connection with device ID ");
     Serial.print(bus.packets[data].content[0], DEC);
