+++ examples/ARDUINO/Network/SoftwareBitBang/SpeedTest/Transmitter/Transmitter.ino
@@ -17,7 +17,7 @@ void setup() {
   Serial.begin(115200);
 }
 
-void error_handler(uint8_t code, uint8_t data, void *custom_pointer) {
+void error_handler(uint8_t code, uint16_t data, void *custom_pointer) {
   if(code == PJON_CONNECTION_LOST) {
     Serial.print("Connection lost with device id ");
     Serial.println(bus.packets[data].content[0], DEC);
